Comentado código y añadida puntiación al tutorial (15 objects)
